pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-research-agent"
version = "0.1.0"
description = "End-to-end automated AI research paper generation system"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "requests",
    "arxiv",
    "docker",
    "python-dotenv",
    "pyyaml",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "ruff",
    "mypy",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
